+++ martin_binance/service/funds_rate_exporter.py
@@ -7,9 +7,11 @@
 __author__ = "Jerry Fedorenko"
 __copyright__ = "Copyright © 2021 Jerry Fedorenko aka VM"
 __license__ = "MIT"
-__version__ = "2.0.4"
+__version__ = "2.0.4b2"
 __maintainer__ = "Jerry Fedorenko"
 __contact__ = 'https://github.com/DogsTailFarmer'
+
+from datetime import datetime
 
 import os
 import time
@@ -118,9 +120,10 @@ def get_rate(_currency_rate) -> {}:
     session = Session()
     session.headers.update(headers)
 
-    for currency in list(_currency_rate.keys()):
+    for currency in _currency_rate:
         _currency = replace.get(currency, currency)
         price = -1
+        '''
         parameters = {'amount': 1, 'symbol': 'USD', 'convert': _currency}
         try:
             response = session.get(URL, params=parameters)
@@ -137,9 +140,9 @@ def get_rate(_currency_rate) -> {}:
             if response.status_code == 200:
                 data = response.json()
                 price = data['data'][0]['quote'][_currency]['price'] or -1
-
+        '''
         _currency_rate[currency] = price
-        time.sleep(request_delay)
+        # time.sleep(request_delay)
     return _currency_rate