+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * `solve()`, `calc_grid()`: use [scipy.optimize.minimize](https://docs.scipy.org/doc/scipy/reference/generated/scipy.optimize.minimize.html#scipy-optimize-minimize)
 for calculation Reverse cycle over price
 * optimize CPU loads at last_state_update()
+* convert_tp(): clear start_hold state and initialize restore grid orders 
 
 ## 2.0.3 - 2023-12-01
 ### Fix