+++ CHANGELOG.md
@@ -1,10 +1,11 @@
-## 2.0.4 - 2023-12-xx
+## 2.0.4 - 2023-12-08
 ### Update
 * `funds_rate_exporter.py`: average CPU load present as percent value at 1 minute period
 * `solve()`, `calc_grid()`: use [scipy.optimize.minimize](https://docs.scipy.org/doc/scipy/reference/generated/scipy.optimize.minimize.html#scipy-optimize-minimize)
 for calculation Reverse cycle over price
 * optimize CPU loads at last_state_update()
-* convert_tp(): clear start_hold state and initialize restore grid orders 
+* convert_tp(): clear start_hold state and initialize restore grid orders
+* Up requirements for exchanges-wrapper==1.4.1
 
 ## 2.0.3 - 2023-12-01
 ### Fix